free-semiring: split the old and new implementations
This splits the old and new implementation of FreeSemiring.  The new
implementation is a drop-in replacement of the old one. By default we
build the new one, however, you can build the old one by calling cmake
with -DOLD_SEMIRING=1.

The commit also includes some changes to the evaluation of
FreeSemiring over some other semiring.  Now the mapping used during
evalutaion maps from VarPtr (instead of from FreeSemiring) to some
semiring value.

Finally this also fixes some minor bugs in the multiplication and star
operations of the new implementation.